+++ freeview/MyCmdLineParser.cpp
@@ -7,8 +7,8 @@
  * Original Author: Ruopeng Wang
  * CVS Revision Info:
  *    $Author: rpwang $
- *    $Date: 2011/04/14 20:30:35 $
- *    $Revision: 1.18 $
+ *    $Date: 2011/12/14 17:13:45 $
+ *    $Revision: 1.19 $
  *
  * Copyright © 2011 The General Hospital Corporation (Boston, MA) "MGH"
  *
@@ -137,7 +137,7 @@ bool MyCmdLineParser::Found( const QString flag, QStringList* args, int nIndex )
   if (ret)
   {
     args->clear();
-    for ( int i = 0; i < sa.size(); i++ )
+    for ( size_t i = 0; i < sa.size(); i++ )
     {
       args->push_back( sa[i].c_str() );
     }
@@ -148,7 +148,7 @@ bool MyCmdLineParser::Found( const QString flag, QStringList* args, int nIndex )
 bool MyCmdLineParser::Found( const char* chFlag, CmdLineEntry* e, int nIndex )
 {
   int n = 0;
-  for ( int i = 0; i < m_cmdLineEntries.size(); i++ )
+  for ( size_t i = 0; i < m_cmdLineEntries.size(); i++ )
   {
     if ( strcmp( m_cmdLineEntries[i].shortName, chFlag ) == 0 ||
          strcmp( m_cmdLineEntries[i].longName, chFlag ) == 0 )
